add `nobrainer.dataset` namespace
