+++ nobrainer/tests/volume_test.py
@@ -6,12 +6,6 @@ import tensorflow as tf
 from nobrainer import volume
 
 
-# TODO: need to implement this soon.
-@pytest.mark.xfail
-def test_get_dataset():
-    assert False
-
-
 def test_apply_random_transform():
     shape = (10, 10, 10)
     x = np.ones(shape).astype(np.float32)
@@ -154,34 +148,3 @@ def test_from_blocks():
     block_shape = (2, 2, 2)
     outputs = volume.from_blocks(volume.to_blocks(x, block_shape), x.shape)
     assert_array_equal(outputs, x)
-
-
-def test_get_steps_per_epoch():
-    nsteps = volume.get_steps_per_epoch(
-        n_volumes=1,
-        volume_shape=(256, 256, 256),
-        block_shape=(64, 64, 64),
-        batch_size=1,
-    )
-    assert nsteps == 64
-    nsteps = volume.get_steps_per_epoch(
-        n_volumes=1,
-        volume_shape=(256, 256, 256),
-        block_shape=(64, 64, 64),
-        batch_size=64,
-    )
-    assert nsteps == 1
-    nsteps = volume.get_steps_per_epoch(
-        n_volumes=1,
-        volume_shape=(256, 256, 256),
-        block_shape=(64, 64, 64),
-        batch_size=63,
-    )
-    assert nsteps == 2
-    nsteps = volume.get_steps_per_epoch(
-        n_volumes=10,
-        volume_shape=(256, 256, 256),
-        block_shape=(128, 128, 128),
-        batch_size=4,
-    )
-    assert nsteps == 20
